Move Mount-Workspace function from device specific module to platform specific module
and rename it to `Mount-RemoteVolume`, because this function is a general function.

Also, this commit implements `Mount-Workspace W4` (@note: `W4` instead of `W4:` is used as workspace identifier, since it is more succinct, and easier to use as JSON key).
There if no need to explicitly call `Mount-RemoteVolume` w/ FULL parameters every time.
An GPG encrypted `~/.ohmypwsh.d/workspace-info.asc` workspace information file should be supplied.